Tidy up mazeHelper
I have decided to just define `struct Node` in the mazeHelper. So every
implementation will use the same `struct`, regardless of the need to use
every member of the `struct`. For example, `backtrack.cpp` doesn't need
the `group` variable, but it's still needed by `kruskall.cpp` so we keep
it there.

Some constant variables have also been changed to be a macro (i.e.
created using the `define` directive).